macosx: compilation fix for the 10.4u SDK and removal of a hack
Note that all NS*Integer types were introduced on 10.5 only. They can tollfree bridged with the respective plain ints. They are undocumented anyway...

+++ modules/gui/macosx/embeddedwindow.h
@@ -53,7 +53,7 @@
     BOOL              b_window_is_invisible;
 
     NSSize videoRatio;
-    NSInteger originalLevel;
+    int originalLevel;
 }
 
 - (void)controlTintChanged;
